Ok it has stopped throwing exceptions now at least.
So the issue was that when you do FillBlob, it does not like being
passed in an uninitialised Mat. So just construct a Mat of the same
size as the frame from the webcam and give it that.

Unfortunately, it thinks the entire frame is a blob, which is not
very helpful. It would help if this library actually had some kind
of documentation.

Will read source code to see if we can fix this.